Lazer to Base (#118)
* feat: add multiple states for countdown component

* feat: teaser page first pass

* feat: iskra config

* feat: implement community page

* chore: remove console.log

* feat: oak config

* feat: include open sea

* feat: fwb x bonfire drop

* feat: fini config

* feat: implement teaser page

* feat: auto connect cb wallet

* feat: modal balance status polling for insufficient funds

* fix: overflow issue on CM form

* chore: reduce size of icon

* feat: ens for address pill

* fix: isBeforeStartDate isAfterStartDate

* feat: handle bridge error

* fix: add zoned times

* feat: use UTC dates in config

* chore: change start and end dates to utc

* feat: quantity button in mint dialog

* feat: tweets loading work complete

* feat: validate mints with limited quantity

* fix: remove console.log

* feat: tabs drop cards using background and icons

* feat: layout for tabs drop card images and clickable area

* feat: twitter card work

* chore: twitter module on community page

* chore: moved to using fetch for twitter call

* chore: hiding component when no tweets are returned

* fix: updated calculation

* fix: base eth -> eth

* chore: upgraded thirdweb sdk and added client id to the provider

* fix: update grants copy

* chore: removed carrot

* chore: merged staging

* chore: added teaser page

* chore: fixed teaser page

* fix: add grant info back to grants page, update header

* fix: teaser page

* fix: crash when no tweets due to rate limit

* chore: hiding onchain fun heading in the navbar between breakpoints to make room for buttons

* chore: changed presentation of startdate and endate on the API with UTC timestamps

* chore: caching ens requests

* fix: not found resource styling

* chore: minor design tweak

* fix: schedule format

* fix: corrected the gas price

* feat: xmtp flow

* fix: used correct icon

* chore: removed unused code

* feat: implement mirror subscription modal and footer input

* chore: design not found page

* chore: added more padding

* fix: stabalized bridge feature

* chore: rounding the amount

* chore: updated default amount

* chore: upgraded parallel icon and styles

* fix: infinite loop in a flow

* chore: optimized the logic

* chore: optimized the logic

* chore: fix mobile menu scroll, update trending page description

* chore: fix scroll on modal, conditionally render buy with cc

* chore: fix button sizing, order of buttons, size of text on teaser page

* Feat: Swap in RainbowKit and WAGMI for ThirdWeb SDK (#25)

* feat: started on using rainbowkit

* fix: fully working, a few polish things needed

* fix: cleanup

* feat: useIsMisMatched

* feat: removing thirdweb

* fix: cleanup

* fix: mobile ios issues

* Update AddressPill.tsx

---------


* chore: updated XMTP flow

* chore: triggering a build with a small change

* fix: added additional fail safe on mintDialog to avoid reassessing balance on every screen

* fix: bridge prefill amount

* fix: rounding, populate used value

* chore: only require password if the ENV is present

* chore: added utc format to object key

* chore: setting max width of XMTP dialog

* chore: add description for hero mint

* chore: moved around fetch call

* chore: change browser theme color

* feat: add new 0308 teaser

* feat: condionally render text in mint button for free mints

* chore: styling changes to XMTP dialog

* chore: add cb partner images

* feat: add 0308 page, add connect wallet teaser nav

* chore: fix spacing

* chore: move address pill above desc

* chore: fix dates and contract address

* chore: minor UI fixes

* chore: insufficient fund without CC flow

* chore: added privacy policy and terms of service footer

* chore: marked subscribe with wallet as external link

* chore: change address pill color in all mint flows

* chore: triggering build

* feat: mint dialog close button mobile

* fix: partner pages live at  redirect old routes to new

* fix: share button not clickable

* chore: qa po round 3

* chore: logos updated

* chore: final touches and nest coin logo

* chore: change back blackbird logo

---------
