+++ src/components/MintDialog/pages/Success/Success.tsx
@@ -9,7 +9,7 @@ import { PartnerInfo } from '../../elements/PartnerInfo'
 import { ViewOnExplorer } from '../../elements/ViewOnExplorer'
 import dialogClasses from '@/components/dialog.module.css'
 import { Address } from 'viem'
-import { Share } from '@/components/Share';
+import { Share } from '@/components/Share'
 
 interface SuccessProps {
   resetModal: () => void
@@ -52,10 +52,10 @@ export const Success: FC<SuccessProps> = ({
         />
       </Dialog.Close>
 
-      <Dialog.DialogContent>
-          <h3 className="font-sans text-[16px] text-[#151515]">Share on </h3>
-          <Share/>
-      </Dialog.DialogContent>
+      <div>
+        <h3 className="font-sans text-[16px] text-[#151515]">Share on </h3>
+        <Share />
+      </div>
     </>
   )
 }
