+++ src/config/partners/blackbird.ts
@@ -8,7 +8,7 @@ const blackbird: Partner = {
   description:
     'Blackbird is a community owned platform for creators to monetize their digital work.',
   brandColor: '#000000',
-  icon: '/partners/blackbird/icon.png',
+  icon: '/partners/blackbird/icon.svg',
   aarweaveDigest: 'GjssNdA6XK7VYynkvwDem3KYwPACSU9nDWpR5rei3hw',
   twitter: '@blackbird_xyz',
   drops: [
