+++ src/config/partners/fwb.ts
@@ -8,7 +8,7 @@ const fwb: Partner = {
   description:
     'Friends With Benefits is a community of builders, creatives, and investors who believe in the power of social tokens and the communities they create. We are a decentralized autonomous organization (DAO) that supports the growth of the social token ecosystem through community grants, educational initiatives, and community events.',
   brandColor: '#000000',
-  icon: '/partners/fwb/icon.jpg',
+  icon: '/partners/fwb/icon.svg',
   aarweaveDigest: 'GjssNdA6XK7VYynkvwDem3KYwPACSU9nDWpR5rei3hw',
   twitter: '@FWBtweets',
   drops: [
